python-requirements.txt
spyne==2.12.10
MySQL-python==1.2.5
Flask==0.10.1
Flask-Babel==0.9
rdflib==4.2.1
skosprovider-rdf==0.4.1
skosprovider-sqlalchemy==0.4.2
alembic==0.8.4
psycopg2==2.6.1
